console: Create abstract ArduinoConsoleBackend
Heavily based on the existing SBArduinoConsoleBackend implementation
