+++ j5/backends/console/j5/__init__.py
@@ -1,0 +1,1 @@
+"""Abstract console backend implementions provided by j5."""
